requirements.txt
Flask==2.0.1
Flask-MySQL==1.5.2
mysql-connector-python==8.0.26
python-dotenv==0.19.0
Werkzeug==2.0.1 